+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log for SharePointDsc
 
+## Unreleased
+
+* Added TermStoreAdministrators property to SPManagedMetadataServiceApp
+
 ## 1.6
 
 * Updated SPWebApplication to allow Claims Authentication configuration